backend/requirements.txt
fastapi
uvicorn
postgrest
pydantic
supabase
python-dotenv
httpx
